requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart>=0.0.9
pypdf>=4.0.0
openai>=1.65.0
httpx>=0.28.0
python-dotenv>=1.0.1
pydantic>=2.10.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
cryptography>=42.0.0
